ml-service/requirements.txt
flask
flask-cors
pdfplumber
python-docx
scikit-learn
nltk
numpy
